LayoutTests/platform/blackberry/fast/forms/validationMessage-expected.txt

Test for validationMessage DOM property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL patternInput.validationMessage should be pattern mismatch. Was Pattern mismatch.
FAIL requiredInput.validationMessage should be value missing. Was Input value missing.
FAIL requiredTextArea.validationMessage should be value missing. Was Input value missing.
FAIL requiredSelect.validationMessage should be value missing. Was Input value missing.
FAIL emailInput.validationMessage should be type mismatch. Was Valid email address required.
PASS but.validationMessage is ''
PASS anoninput.validationMessage is ''
PASS happyFieldset.validationMessage is ''
PASS happySelect.validationMessage is ''
PASS happyOutput.validationMessage is ''
PASS happyObject.validationMessage is ''
PASS happyKeygen.validationMessage is ''
PASS successfullyParsed is true

TEST COMPLETE
